requirements.txt
torch>=2.3.0
transformers>=4.45.0
datasets>=2.18.0
flower>=1.8.0
hydra-core>=1.3.2
bitsandbytes>=0.45.0
accelerate>=0.30.0
peft>=0.10.0
sentencepiece
protobuf
wandb>=0.16.0
omegaconf
flwr>=1.11.0
flwr[simulation]>=1.11.0  # optional but useful for local multi-client sim
codecarbon>=2.7.0
